Lab11
When coding and thinking about the mobile, I know that for next time the design should be stacked on top of each other for easier accessibility for the page! I tried to get as close to my mobile design as possible, but it was challenging to get things lined up. If I were to redesign, I would make sure to keep that in mind.